refactor: 휴대폰인증, 온보딩, 회원 정보 조회 APi 응답 DTO를 UserResponse로 통합 (#82)
