--- src/main/java/com/team/buddyya/certification/dto/response/VerifyCodeResponse.java
@@ -1,74 +1,0 @@
-package com.team.buddyya.certification.dto.response;
-
-import com.team.buddyya.student.domain.Student;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-public record VerifyCodeResponse(
-        Long id,
-        String name,
-        String country,
-        String university,
-        String gender,
-        String profileImageUrl,
-        Boolean isCertificated,
-        Boolean isStudentIdCardRequested,
-        Boolean isKorean,
-        List<String> majors,
-        List<String> languages,
-        List<String> interests,
-        String phoneNumber,
-        String status,
-        String accessToken,
-        String refreshToken) {
-
-    public static VerifyCodeResponse from(Student student, Boolean isStudentIdCardRequested, String phoneNumber, String status, String accessToken, String refreshToken) {
-        return new VerifyCodeResponse(
-                student.getId(),
-                student.getName(),
-                student.getCountry(),
-                student.getUniversity().getUniversityName(),
-                student.getGender().getDisplayName(),
-                student.getProfileImage().getUrl(),
-                student.getIsCertificated(),
-                isStudentIdCardRequested,
-                student.getIsKorean(),
-                convertToStringList(student.getMajors()),
-                convertToStringList(student.getLanguages()),
-                convertToStringList(student.getInterests()),
-                phoneNumber,
-                status,
-                accessToken,
-                refreshToken
-        );
-    }
-
-    public static VerifyCodeResponse fromNewMember(String phoneNumber, String status) {
-        return new VerifyCodeResponse(
-                null,  // id
-                null,  // name
-                null,  // country
-                null,  // university
-                null,  // gender
-                null,  // profileImageUrl
-                false, // isCertificated
-                false, // isStudentIdCardRequested
-                null,  // isKorean
-                null,  // majors
-                null,  // languages
-                null,  // interests
-                phoneNumber,
-                status,
-                null,  // accessToken
-                null   // refreshToken
-        );
-    }
-
-    private static List<String> convertToStringList(List<?> list) {
-        return list.stream()
-                .map(Object::toString)
-                .collect(Collectors.toList());
-    }
-}
-
