+++ src/main/java/com/team/buddyya/common/DataLoader.java
@@ -10,6 +10,7 @@ import com.team.buddyya.student.controller.UserController;
 import com.team.buddyya.student.domain.*;
 import com.team.buddyya.student.dto.request.OnBoardingRequest;
 import com.team.buddyya.student.dto.response.OnBoardingResponse;
+import com.team.buddyya.student.dto.response.UserResponse;
 import com.team.buddyya.student.repository.*;
 
 import java.util.List;
@@ -115,7 +116,7 @@ public class DataLoader implements CommandLineRunner {
                 )
         );
         mockRequests.forEach(request -> {
-            ResponseEntity<OnBoardingResponse> response = userController.onboard(request);
+            ResponseEntity<UserResponse> response = userController.onboard(request);
         });
     }
